+++ usb-hid-s3/test/test_device_identity/test_device_identity.cpp
@@ -15,8 +15,8 @@ void test_device_id_from_mac() {
 
 void test_soft_ap_ssid_uses_friendly_mac_name() {
   char ssid[24];
-  formatSoftApSsid("de94", ssid);
-  TEST_ASSERT_EQUAL_STRING("InputPilot-Dupe9", ssid);
+  formatSoftApSsid("aabbccddde94", ssid);
+  TEST_ASSERT_EQUAL_STRING("InputPilot-Halo9", ssid);
 }
 
 void test_mdns_hostname_lower_suffix() {
@@ -27,17 +27,18 @@ void test_mdns_hostname_lower_suffix() {
 
 void test_device_name_matches_soft_ap_name() {
   char name[24];
-  formatDeviceName("DE94", name);
-  TEST_ASSERT_EQUAL_STRING("InputPilot-Dupe9", name);
+  formatDeviceName("AABBCCDDDE94", name);
+  TEST_ASSERT_EQUAL_STRING("InputPilot-Halo9", name);
 }
 
-void test_friendly_name_has_deterministic_fallbacks() {
-  char digitsOnly[24];
-  char lettersOnly[24];
-  formatDeviceName("1234", digitsOnly);
-  formatDeviceName("beef", lettersOnly);
-  TEST_ASSERT_EQUAL_STRING("InputPilot-Bolt1", digitsOnly);
-  TEST_ASSERT_EQUAL_STRING("InputPilot-Bolt1", lettersOnly);
+void test_friendly_name_uses_complete_mac_and_first_digit() {
+  char first[24];
+  char second[24];
+  formatDeviceName("aabbccddde94", first);
+  formatDeviceName("11223344de94", second);
+  TEST_ASSERT_NOT_EQUAL(0, strcmp(first, second));
+  TEST_ASSERT_EQUAL('9', first[strlen(first) - 1]);
+  TEST_ASSERT_EQUAL('1', second[strlen(second) - 1]);
 }
 
 void test_null_mac_rejected() {
@@ -55,7 +56,7 @@ int main() {
   RUN_TEST(test_soft_ap_ssid_uses_friendly_mac_name);
   RUN_TEST(test_mdns_hostname_lower_suffix);
   RUN_TEST(test_device_name_matches_soft_ap_name);
-  RUN_TEST(test_friendly_name_has_deterministic_fallbacks);
+  RUN_TEST(test_friendly_name_uses_complete_mac_and_first_digit);
   RUN_TEST(test_null_mac_rejected);
   return UNITY_END();
 }